// src/firebase/dto/user.dto.ts
export class UserDto {
  userId: string;
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  birthDate: Date;
  accountStatus: string;
  isEmailVerified: boolean;
  isVerified: boolean;
  subscriptionPlan: string;
  assignedLocker: string;
  walletReference: string;
  branchReference: string;
  photo: string;
} 